Add Edit Configuration
Refactor ConfigurationForm into its own component

Optionally pass in an initial value to the form

Use the result from the form to either append or edit the
Configurations list
