+++ src/lib/ConfigType.js
@@ -1,0 +1,8 @@
+const ConfigType = {
+  SWITCH: 0,
+  STRING: 1,
+  NUMBER: 2,
+};
+Object.freeze(ConfigType);
+
+export default ConfigType;
